bpmn2: added new base test class which should fix the coloring issue on jenkins (currently the job is red even though no errors where present)
