+++ tests/org.jboss.tools.bpmn2.ui.bot.test/src/org/jboss/tools/bpmn2/ui/bot/test/testcase/editor/AdHocProcessTest.java
@@ -9,7 +9,6 @@ import org.jboss.tools.bpmn2.reddeer.editor.jbpm.gateways.ExclusiveGateway;
 import org.jboss.tools.bpmn2.reddeer.editor.jbpm.startevents.StartEvent;
 import org.jboss.tools.bpmn2.ui.bot.test.JBPM6BaseTest;
 import org.jboss.tools.bpmn2.ui.bot.test.requirements.ProcessDefinitionRequirement.ProcessDefinition;
-import org.junit.Test;
 
 /**
  * ISSUES:
@@ -21,12 +20,8 @@ import org.junit.Test;
 @ProcessDefinition(name="BPMN2-AdHocProcess",  project="EditorTestProject")
 public class AdHocProcessTest extends JBPM6BaseTest {
 
-	/**
-	 *  
-	 * @throws Exception
-	 */
-	@Test
-	public void runTest() throws Exception {
+	@Override
+	public void buildProcessModel() {
 		BPMN2Process process = new BPMN2Process("BPMN2-AdHocProcess");
 		process.setAddHoc(true);
 		process.add("Task 3", ConstructType.SCRIPT_TASK);
