+++ tests/org.jboss.tools.bpmn2.ui.bot.test/src/org/jboss/tools/bpmn2/ui/bot/test/testcase/editor/BooleanStructureReferenceTest.java
@@ -10,7 +10,6 @@ import org.jboss.tools.bpmn2.reddeer.editor.jbpm.activities.UserTask;
 import org.jboss.tools.bpmn2.reddeer.editor.jbpm.startevents.StartEvent;
 import org.jboss.tools.bpmn2.ui.bot.test.JBPM6BaseTest;
 import org.jboss.tools.bpmn2.ui.bot.test.requirements.ProcessDefinitionRequirement.ProcessDefinition;
-import org.junit.Test;
 
 /**
  *     
@@ -19,8 +18,8 @@ import org.junit.Test;
 @ProcessDefinition(name="BPMN2-BooleanStructureRef", project="EditorTestProject")
 public class BooleanStructureReferenceTest extends JBPM6BaseTest {
 
-	@Test
-	public void runTest() throws Exception {
+	@Override
+	public void buildProcessModel() {
 		BPMN2Process process = new BPMN2Process("BPMN2-BooleanStructureRef");
 		process.addDataType("Boolean");
 		process.addLocalVariable("test", "Boolean");
