+++ tests/org.jboss.tools.bpmn2.ui.bot.test/src/org/jboss/tools/bpmn2/ui/bot/test/testcase/editor/BoundaryEscalationEventOnTaskTest.java
@@ -10,7 +10,6 @@ import org.jboss.tools.bpmn2.reddeer.editor.jbpm.gateways.ParallelGateway;
 import org.jboss.tools.bpmn2.reddeer.editor.jbpm.startevents.StartEvent;
 import org.jboss.tools.bpmn2.ui.bot.test.JBPM6BaseTest;
 import org.jboss.tools.bpmn2.ui.bot.test.requirements.ProcessDefinitionRequirement.ProcessDefinition;
-import org.junit.Test;
 
 /**
  * ISSUE: language should be 'http://www.jboss.org/drools/rule' but it's not available.
@@ -20,8 +19,8 @@ import org.junit.Test;
 @ProcessDefinition(name="BPMN2-BoundaryEscalationEventOnTask", project="EditorTestProject")
 public class BoundaryEscalationEventOnTaskTest extends JBPM6BaseTest {
 
-	@Test
-	public void runTest() throws Exception {
+	@Override
+	public void buildProcessModel() {
 		StartEvent startEvent = new StartEvent("StartProcess");
 		startEvent.append("Split", ConstructType.PARALLEL_GATEWAY);
