+++ tests/org.jboss.tools.bpmn2.ui.bot.test/src/org/jboss/tools/bpmn2/ui/bot/test/testcase/editor/CallActivityTest.java
@@ -12,7 +12,6 @@ import org.jboss.tools.bpmn2.reddeer.editor.jbpm.activities.CallActivity;
 import org.jboss.tools.bpmn2.reddeer.editor.jbpm.startevents.StartEvent;
 import org.jboss.tools.bpmn2.ui.bot.test.JBPM6BaseTest;
 import org.jboss.tools.bpmn2.ui.bot.test.requirements.ProcessDefinitionRequirement.ProcessDefinition;
-import org.junit.Test;
 
 /**
  *     
@@ -21,13 +20,8 @@ import org.junit.Test;
 @ProcessDefinition(name="BPMN2-CallActivity", project="EditorTestProject")
 public class CallActivityTest extends JBPM6BaseTest {
 
-	/**
-	 *
-	 *  
-	 * @throws Exception
-	 */
-	@Test
-	public void runTest() throws Exception {
+	@Override
+	public void buildProcessModel() {
 		BPMN2Process process = new BPMN2Process("BPMN2-CallActivity");
 		process.addDataType("String");
 		process.addLocalVariable("x", "String");
