+++ tests/org.jboss.tools.bpmn2.ui.bot.test/src/org/jboss/tools/bpmn2/ui/bot/test/testcase/editor/ConditionalBoundaryEventInterruptingTest.java
@@ -10,7 +10,6 @@ import org.jboss.tools.bpmn2.reddeer.editor.jbpm.endevents.EscalationEndEvent;
 import org.jboss.tools.bpmn2.reddeer.editor.jbpm.startevents.StartEvent;
 import org.jboss.tools.bpmn2.ui.bot.test.JBPM6BaseTest;
 import org.jboss.tools.bpmn2.ui.bot.test.requirements.ProcessDefinitionRequirement.ProcessDefinition;
-import org.junit.Test;
 
 /**
  *     
@@ -19,13 +18,8 @@ import org.junit.Test;
 @ProcessDefinition(name="BPMN2-ConditionalBoundaryEventInterrupting", project="EditorTestProject")
 public class ConditionalBoundaryEventInterruptingTest extends JBPM6BaseTest {
 
-	/**
-	 *
-	 *  
-	 * @throws Exception
-	 */
-	@Test
-	public void runTest() throws Exception {
+	@Override
+	public void buildProcessModel() {
 		StartEvent startEvent = new StartEvent("StartProcess");
 		startEvent.append("Hello", ConstructType.SUB_PROCESS);
