+++ tests/org.jboss.tools.bpmn2.ui.bot.test/src/org/jboss/tools/bpmn2/ui/bot/test/testcase/editor/ConditionalStartTest.java
@@ -7,7 +7,6 @@ import org.jboss.tools.bpmn2.reddeer.editor.jbpm.startevents.ConditionalStartEve
 import org.jboss.tools.bpmn2.reddeer.editor.jbpm.startevents.StartEvent;
 import org.jboss.tools.bpmn2.ui.bot.test.JBPM6BaseTest;
 import org.jboss.tools.bpmn2.ui.bot.test.requirements.ProcessDefinitionRequirement.ProcessDefinition;
-import org.junit.Test;
 
 /**
  *     
@@ -16,13 +15,8 @@ import org.junit.Test;
 @ProcessDefinition(name="BPMN2-ConditionalStart", project="EditorTestProject")
 public class ConditionalStartTest extends JBPM6BaseTest {
 
-	/**
-	 *
-	 *  
-	 * @throws Exception
-	 */
-	@Test
-	public void runTest() throws Exception {
+	@Override
+	public void buildProcessModel() {
 		new StartEvent("StartProcess").delete();
 		new BPMN2Process("BPMN2-ConditionalStart").add("StartProcess", ConstructType.CONDITIONAL_START_EVENT);
