+++ tests/org.jboss.tools.bpmn2.ui.bot.test/src/org/jboss/tools/bpmn2/ui/bot/test/testcase/editor/DataObjectTest.java
@@ -9,7 +9,6 @@ import org.jboss.tools.bpmn2.reddeer.editor.jbpm.dataobjects.DataObject;
 import org.jboss.tools.bpmn2.reddeer.editor.jbpm.startevents.StartEvent;
 import org.jboss.tools.bpmn2.ui.bot.test.JBPM6BaseTest;
 import org.jboss.tools.bpmn2.ui.bot.test.requirements.ProcessDefinitionRequirement.ProcessDefinition;
-import org.junit.Test;
 
 /**
  *     
@@ -18,13 +17,8 @@ import org.junit.Test;
 @ProcessDefinition(name="BPMN2-DataObject", project="EditorTestProject")
 public class DataObjectTest extends JBPM6BaseTest {
 
-	/**
-	 *
-	 *  
-	 * @throws Exception
-	 */
-	@Test
-	public void runTest() throws Exception {
+	@Override
+	public void buildProcessModel() {
 		StartEvent start = new StartEvent("StartProcess");
 		start.append("Script", ConstructType.SCRIPT_TASK);
