+++ tests/org.jboss.tools.bpmn2.ui.bot.test/src/org/jboss/tools/bpmn2/ui/bot/test/testcase/editor/ErrorEndEventTest.java
@@ -5,7 +5,6 @@ import org.jboss.tools.bpmn2.reddeer.editor.jbpm.endevents.ErrorEndEvent;
 import org.jboss.tools.bpmn2.reddeer.editor.jbpm.startevents.StartEvent;
 import org.jboss.tools.bpmn2.ui.bot.test.JBPM6BaseTest;
 import org.jboss.tools.bpmn2.ui.bot.test.requirements.ProcessDefinitionRequirement.ProcessDefinition;
-import org.junit.Test;
 
 /**
  *     
@@ -14,13 +13,8 @@ import org.junit.Test;
 @ProcessDefinition(name="BPMN2-ErrorEndEvent", project="EditorTestProject")
 public class ErrorEndEventTest extends JBPM6BaseTest {
 
-	/**
-	 *
-	 *  
-	 * @throws Exception
-	 */
-	@Test
-	public void runTest() throws Exception {
+	@Override
+	public void buildProcessModel() {
 		StartEvent start = new StartEvent("StartProcess");
 		start.append("ErrorEvent", ConstructType.ERROR_END_EVENT);
