+++ tests/org.jboss.tools.bpmn2.ui.bot.test/src/org/jboss/tools/bpmn2/ui/bot/test/testcase/editor/InclusiveSplitTest.java
@@ -8,7 +8,6 @@ import org.jboss.tools.bpmn2.reddeer.editor.jbpm.gateways.InclusiveGateway;
 import org.jboss.tools.bpmn2.reddeer.editor.jbpm.startevents.StartEvent;
 import org.jboss.tools.bpmn2.ui.bot.test.JBPM6BaseTest;
 import org.jboss.tools.bpmn2.ui.bot.test.requirements.ProcessDefinitionRequirement.ProcessDefinition;
-import org.junit.Test;
 
 /**
  *     
@@ -17,13 +16,8 @@ import org.junit.Test;
 @ProcessDefinition(name="BPMN2-InclusiveSplit", project="EditorTestProject")
 public class InclusiveSplitTest extends JBPM6BaseTest {
 
-	/**
-	 *
-	 *  
-	 * @throws Exception
-	 */
-	@Test
-	public void runTest() throws Exception {
+	@Override
+	public void buildProcessModel() {
 		BPMN2Process process = new BPMN2Process("BPMN2-InclusiveSplit");
 		process.addDataType("Integer");
 		process.addLocalVariable("x", "Integer");
