+++ tests/org.jboss.tools.bpmn2.ui.bot.test/src/org/jboss/tools/bpmn2/ui/bot/test/testcase/editor/MultipleStartEventTest.java
@@ -10,7 +10,6 @@ import org.jboss.tools.bpmn2.reddeer.editor.jbpm.startevents.TimerStartEvent;
 import org.jboss.tools.bpmn2.reddeer.editor.jbpm.startevents.TimerStartEvent.Type;
 import org.jboss.tools.bpmn2.ui.bot.test.JBPM6BaseTest;
 import org.jboss.tools.bpmn2.ui.bot.test.requirements.ProcessDefinitionRequirement.ProcessDefinition;
-import org.junit.Test;
 
 /**
  *     
@@ -19,13 +18,8 @@ import org.junit.Test;
 @ProcessDefinition(name="BPMN2-MultipleStartEvent", project="EditorTestProject")
 public class MultipleStartEventTest extends JBPM6BaseTest {
 
-	/**
-	 *
-	 *  
-	 * @throws Exception
-	 */
-	@Test
-	public void runTest() throws Exception {
+	@Override
+	public void buildProcessModel() {
 		StartEvent start = new StartEvent("StartProcess");
 		start.append("StartTimer", ConstructType.TIMER_START_EVENT, Position.SOUTH);
 		start.append("Split", ConstructType.EXCLUSIVE_GATEWAY, Position.SOUTH_EAST);
