+++ tests/org.jboss.tools.bpmn2.ui.bot.test/src/org/jboss/tools/bpmn2/ui/bot/test/testcase/editor/OnEntryExitMixedNamespacedScriptProcessTest.java
@@ -6,7 +6,6 @@ import org.jboss.tools.bpmn2.reddeer.editor.jbpm.activities.Task;
 import org.jboss.tools.bpmn2.reddeer.editor.jbpm.startevents.StartEvent;
 import org.jboss.tools.bpmn2.ui.bot.test.JBPM6BaseTest;
 import org.jboss.tools.bpmn2.ui.bot.test.requirements.ProcessDefinitionRequirement.ProcessDefinition;
-import org.junit.Test;
 
 /**
  *     
@@ -15,13 +14,8 @@ import org.junit.Test;
 @ProcessDefinition(name="BPMN2-OnEntryExitMixedNamespacedScriptProcess", project="EditorTestProject")
 public class OnEntryExitMixedNamespacedScriptProcessTest extends JBPM6BaseTest {
 
-	/**
-	 *
-	 *  
-	 * @throws Exception
-	 */
-	@Test
-	public void runTest() throws Exception {
+	@Override
+	public void buildProcessModel() {
 		BPMN2Process process = new BPMN2Process("BPMN2-OnEntryExitMixedNamespacedScriptProcess");
 		process.addImport("java.util.List");
